rm duplication
